requirements.txt
django
djangorestframework
psycopg2-binary
requests
pymed
gunicorn
daphne
celery
redis
torch
monai
numpy
python-decouple
django-cors-headers
pillow
django-redis
channels
channels-redis
netaddr
whitenoise
django-filter
markdown
djangorestframework-simplejwt
drf-yasg
pytest
pytest-django
coverage
factory-boy
faker
black
isort
flake8
mypy
